feat: 댓글 조회에 좋아요, 싫어요 추가

+++ src/main/java/com/suhwan/cowtalk/comment/repository/CommentGoodBadRepository.java
@@ -10,5 +10,5 @@ public interface CommentGoodBadRepository extends JpaRepository<CommentGoodBad, 
 
   boolean existsByCommentAndMember(Comment comment, Member member);
 
-  int countByCommentAndGoodBad(Comment comment, GoodBad goodBad);
+  long countByCommentAndGoodBad(Comment comment, GoodBad goodBad);
 }
